+++ docs/CHANGELOG.md
@@ -54,3 +54,4 @@
 - Updated comments on <fc-select>.
 - Removed onFocus listener from <fc-combobox> and switched to onClick for better UX.
 - Updated onOptionSelect to prevent element focus loss when clicking on an option.
+- Fixed <fc-option> checked-icon slot sizing.